Implement using StateT
This version has the nice property that we don't need to thread
the game identifiers from one call of makePick to the next in
the makePicks function. Instead the list is maintened inside
the state. The threading and mutating of the state is now handled
by selectGame which gets the current state, uses it for a selection,
and then puts back the mutated state.